Update README.md
aaayafuj_hack (Yafet Yohanes)

This is a cybersecurity tool designed to perform phishing attacks on a variety of platforms, developed by Yafet Yohanes. The tool is optimized for use on Kali Linux and Termux environments and can be easily installed by anyone interested in cybersecurity research and ethical hacking.

Key features include:

Phishing Attack Selection: Choose from a variety of phishing templates and attack methods.
User-Friendly Interface: Simple commands and straightforward setup.
Open Source: Fully available for modification and improvements by the community.
Disclaimer: This project is intended for educational purposes only. The user is responsible for ensuring that the tool is used legally and ethically.